feat: improve creating keys
