+++ packages/ui5-cc-spreadsheetimporter/src/controller/odata/MetadataHandler.ts
@@ -37,4 +37,5 @@ export default abstract class MetadataHandler extends ManagedObject {
 	abstract getLabelList(columns: Columns, odataType: string, odataEntityType: any, excludeColumns: Columns): ListObject;
 	abstract getKeyList(odataEntityType: any): string[];
 	abstract getODataEntitiesRecursive(entityName: string, deepLevel: number): any;
+	abstract getKeys(binding: any, payload: any, IsActiveEntity?: boolean, excludeIsActiveEntity?: boolean): Record<string, any>;
 }
